+++ src/londonsw/controller/StartUpController.java
@@ -30,6 +30,7 @@ import londonsw.view.simulation.VehicleGUIDecorator;
 //import java.awt.*;
 import java.io.IOException;
 
+
 /**
  *
  */
@@ -43,8 +44,6 @@ public class StartUpController extends Application{
     @FXML private TextField width;
 
     @FXML private TextField height;
-
-    @FXML public  MapLoadingController main;
 
     @FXML public static String mapName;
 
@@ -86,13 +85,11 @@ public class StartUpController extends Application{
         chooser.setTitle("Open File");
         File file = chooser.showOpenDialog(new Stage());
 
-        mapName=file.getName();
+        mapName=file.getName().toString();
 
         if(file!=null)
         {
             Map map = Map.loadMap(file.getName());
-
-            // Map map = new Map(20,20);
 
             //Decorate map to extend to GUI functionality
             MapGridGUIDecorator mapGridGUIDecorator = new MapGridGUIDecorator(map.getGrid());
@@ -104,12 +101,12 @@ public class StartUpController extends Application{
             //Instantiate GridPane that will contain empty map with grass
             GridPane root = mapGridGUIDecorator.drawComponents();
 
-
             p.getChildren().add(root);
 
             Stage stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
 
             stage.setScene(new Scene(simulationModeScreen));
+            stage.setFullScreen(true);
         }
         else
         {
@@ -152,10 +149,62 @@ public class StartUpController extends Application{
         Stage stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
 
         stage.setScene(new Scene(mapCreation));
+
+        stage.setFullScreen(true);
+    }
+    public void StartSimulation(ActionEvent actionEvent) throws Exception {
+
+        Map map=Map.loadMap(mapName);
+        Parent simulationModeScreen  = FXMLLoader.load(getClass().getResource("../view/startup/SimulationMode" + ".fxml"));
+        Node node = simulationModeScreen.lookup("#Scene");
+
+        Pane p = (Pane) node;
+
+
+        //Decorate map to extend to GUI functionality
+        MapGridGUIDecorator mapGridGUIDecorator = new MapGridGUIDecorator(map.getGrid());
+
+        //Always apply resize
+        mapGridGUIDecorator.setResizeFactor(new ResizeFactor(5.0/map.getWidth(),5.0/map.getHeight()));
+
+
+        //Instantiate GridPane that will contain empty map with grass
+        GridPane root = mapGridGUIDecorator.drawComponents();
+
+        p.getChildren().add(root);
+
+
+        Stage stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
+
+        stage.setScene(new Scene(simulationModeScreen));
+
     }
 
-    public void init(MapLoadingController mapLoadingController) {
-        main=mapLoadingController;
+    public void resetSimulation(ActionEvent actionEvent)throws Exception {
+        Parent simulationModeScreen  = FXMLLoader.load(getClass().getResource("../view/startup/SimulationMode" + ".fxml"));
+
+        Node node = simulationModeScreen.lookup("#Scene");
+        Pane p = (Pane) node;
+
+        //Create map
+        Map map = new Map(20,20);
+
+        //Decorate map to extend to GUI functionality
+        MapGridGUIDecorator mapGridGUIDecorator = new MapGridGUIDecorator(map.getGrid());
+
+        //Always apply resize
+        mapGridGUIDecorator.setResizeFactor(new ResizeFactor(5.0/map.getWidth(),5.0/map.getHeight()));
+
+
+        //Instantiate GridPane that will contain empty map with grass
+        GridPane root = mapGridGUIDecorator.drawComponents();
+
+
+        p.getChildren().add(root);
+
+        Stage stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
+
+        stage.setScene(new Scene(simulationModeScreen));
     }
 
 }
